fix(OntologyBrowser): #2529 comments review
